Rewritten for buffered reading instead of using an memory mapped file. After some very hasty tests this appears to cut memory usage by 50% and runtime by 30% but... SOMEHOW ... this is still slower and less memory efficient than the current Python implementation.
I'm honestly shocked.

+++ aquatx/srna/fast-path/collapser.h
@@ -5,7 +5,7 @@
 #ifndef COLLAPSER_COLLAPSER_H
 #define COLLAPSER_COLLAPSER_H
 
-uintmax_t sequence_counter(char*);
+int sequence_counter(char*);
 
 struct stat statbuf;
 struct c_string_comparator {
